requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.29
psycopg[binary]==3.1.19
psycopg2-binary==2.9.9
pydantic==2.11.9
pandas==2.2.2
numpy==1.26.4
scipy==1.15.3
statsmodels==0.14.2
Jinja2==3.1.4
python-multipart==0.0.9
requests==2.32.3
python-dotenv==1.0.1
httptools==0.6.4
uvloop==0.21.0
websockets==15.0.1
typing-extensions==4.15.0
starlette==0.36.3
tzdata==2025.2
pytz==2025.2
# For auto.arima
pmdarima==2.0.4
scikit-learn==1.4.2
rq==1.16.2
redis==5.0.4
